Draw real icons or empty
