+++ Zilon.Core/CDT.LAST.MonoGameClient/Content/Content.mgcb
@@ -579,3 +579,14 @@
 /processorParam:TextureFormat=Color
 /build:Sprites/ui/walk-cursor.png
 
+#begin Sprites/ui/EmptyPropIcon.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:Sprites/ui/EmptyPropIcon.png
